Add more tests for screenshots and fix existing tests
